feat: Add beatmap events; feat: Add events to database on hype/set status/clear hype; feat: Add api methods to check beatmap events;

+++ Sunrise.Shared/Database/Services/Beatmaps/CustomBeatmapStatusService.cs
@@ -5,7 +5,7 @@ using Sunrise.Shared.Database.Extensions;
 using Sunrise.Shared.Database.Models.Beatmap;
 using Sunrise.Shared.Database.Objects;
 using Sunrise.Shared.Database.Repositories;
-using Sunrise.Shared.Utils;
+using Sunrise.Shared.Database.Services.Events;
 
 namespace Sunrise.Shared.Database.Services.Beatmaps;
 
@@ -13,7 +13,8 @@ public class CustomBeatmapStatusService(
     ILogger<ScoreRepository> logger,
     Lazy<DatabaseService> databaseService,
     SunriseDbContext dbContext,
-    BeatmapHypeService beatmapHypeService)
+    BeatmapHypeService beatmapHypeService,
+    BeatmapEventService beatmapEventService)
 {
     public async Task<CustomBeatmapStatus?> GetCustomBeatmapStatus(string beatmapHash, QueryOptions? options = null, CancellationToken ct = default)
     {
@@ -35,9 +36,13 @@ public class CustomBeatmapStatusService(
     {
         return await databaseService.Value.CommitAsTransactionAsync(async () =>
         {
-            var applyBeatmapHypesResult = await beatmapHypeService.ApplyBeatmapHypes(status.BeatmapSetId, status.Status);
+            var applyBeatmapHypesResult = await beatmapHypeService.ApplyBeatmapHypes(status.UpdatedByUserId, status.BeatmapSetId, status.Status);
             if (applyBeatmapHypesResult.IsFailure)
                 throw new ApplicationException(applyBeatmapHypesResult.Error);
+
+            var addBeatmapStatusChangedEventResult = await beatmapEventService.AddBeatmapStatusChangedEvent(status.UpdatedByUserId, status.BeatmapSetId, status.BeatmapHash, status.Status);
+            if (addBeatmapStatusChangedEventResult.IsFailure)
+                throw new ApplicationException(addBeatmapStatusChangedEventResult.Error);
 
             dbContext.CustomBeatmapStatuses.Add(status);
             await dbContext.SaveChangesAsync();
@@ -46,8 +51,12 @@ public class CustomBeatmapStatusService(
 
     public async Task<Result> UpdateCustomBeatmapStatus(CustomBeatmapStatus status)
     {
-        return await ResultUtil.TryExecuteAsync(async () =>
+        return await databaseService.Value.CommitAsTransactionAsync(async () =>
         {
+            var addBeatmapStatusChangedEventResult = await beatmapEventService.AddBeatmapStatusChangedEvent(status.UpdatedByUserId, status.BeatmapSetId, status.BeatmapHash, status.Status);
+            if (addBeatmapStatusChangedEventResult.IsFailure)
+                throw new ApplicationException(addBeatmapStatusChangedEventResult.Error);
+
             dbContext.UpdateEntity(status);
             await dbContext.SaveChangesAsync();
         });
@@ -55,8 +64,12 @@ public class CustomBeatmapStatusService(
 
     public async Task<Result> DeleteCustomBeatmapStatus(CustomBeatmapStatus status)
     {
-        return await ResultUtil.TryExecuteAsync(async () =>
+        return await databaseService.Value.CommitAsTransactionAsync(async () =>
         {
+            var addBeatmapStatusChangedEventResult = await beatmapEventService.AddBeatmapStatusChangedEvent(status.UpdatedByUserId, status.BeatmapSetId, status.BeatmapHash, null);
+            if (addBeatmapStatusChangedEventResult.IsFailure)
+                throw new ApplicationException(addBeatmapStatusChangedEventResult.Error);
+
             dbContext.CustomBeatmapStatuses.Remove(status);
             await dbContext.SaveChangesAsync();
         });
